fix(nix-direnv/figlet): Prevent figlet from hanging when the `COLUMNS` env variable is not set
After upgrading nix-direnv to 2.4.0 from 2.2.1 (though 2.3.0 likely also had the
same issue), nix-direnv started hanging in the case it loading an already cached
dev shell. After investigation, the problem turned out to be the call to
`figlet`, which appears in the `shellHook` of our default shell:

    figlet -w$COLUMNS "${flake.description}"

where `${flake.description}` is not imporant - any non-empty string will do.

Assuming a terminal width of 120, normally, the call to `figlet` looks like
this:

    figlet -w120 "Some message"

However, when nix-direnv uses a cached dev shell, the `COLUMNS` is not set and
the call became this:

    figet -w "Some message"

This caused "Some message" to be interpretted as a parameter to the output width
option (`-w`), which turn meant that figlet started waiting for input from
stdin, which of course never comes since nix-direnv executes the dev shell in a
non-interactive sub-shell.