refactor(flake/outputs): Make `packages` contain only the derivation defined in this repo
Previously, the `packages` output contained both packages defined in this repo
as well as re-exported pacakges from third-party flakes. The reason for the
latter is to allow us to easily ensure that those third-party packages build
succefully and get uploaded to our cache.

This is not necessary, as `packages` can simply contain the packages defined by
this repo and `checks` can contain `packages` + re-exported packages.