feat(flake.nix/inputs): Temporary switch from `nixos-24.11` to `staging-next-24.11`
`curl` v8.11.0 was released with a bug that causes `git fetch` (and many
similar commands, including Nix's [fetchers][3]) to fail due to failure to parse
`.netrc` files. Luckily, after the bug was [reported upstream][1], a fix was quickly
proposed and merged by its main [developer][2]. Not long after, the upstream
`curl` patch was [backported][5] to the Nixpkgs `curl` derivation.

Given that we rely on `.netrc` files this bug is a showstopper for us, so its
important that we switch to the patched `curl` version ASAP. Instead of [waiting
for the build to land][5] on the `nixos-24.11` branch, we will switch directly
to the `staging-next-24.11` branch, which already includes the fix.

[1]: https://github.com/curl/curl/issues/15496
[2]: https://github.com/curl/curl/commit/f5c616930b5cf148b1b2632da4f5963ff48bdf88
[3]: https://github.com/NixOS/nixpkgs/issues/356114
[4]: https://github.com/NixOS/nixpkgs/pull/356133
[5]: https://nixpk.gs/pr-tracker.html?pr=356660